Create and provide base url for clickable links `drawDependencies` so we can link the graphs correctly.
+ If no repo url is configured, don't add links
+ Add links to graphs, not to modules
+ Move functions to `CreateGraph.kt`
