+++ plugin-build/plugin/src/main/kotlin/net/c306/dependencygraph/plugin/DependencyGraphTask.kt
@@ -1,8 +1,6 @@
 package net.c306.dependencygraph.plugin
 
 import org.gradle.api.DefaultTask
-import org.gradle.api.Project
-import org.gradle.api.artifacts.ProjectDependency
 import org.gradle.api.plugins.BasePlugin
 import org.gradle.api.provider.ListProperty
 import org.gradle.api.provider.Property
@@ -10,8 +8,6 @@ import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.Optional
 import org.gradle.api.tasks.TaskAction
 import org.gradle.api.tasks.options.Option
-import java.io.File
-import java.util.*
 
 enum class Direction {
     TopToBottom, BottomToTop, LeftToRight, RightToLeft
@@ -61,17 +57,26 @@ abstract class DependencyGraphTask : DefaultTask() {
     abstract val repoRootUrl: Property<String>
 
     @get:Input
-    @get:Option(option = "mainBranchName", description = "Name of the main branch. `main` is used if not provided.")
+    @get:Option(
+        option = "mainBranchName",
+        description = "Name of the main branch. `main` is used if not provided.",
+    )
     @get:Optional
     abstract val mainBranchName: Property<String>
 
     @get:Input
-    @get:Option(option = "graphFileName", description = "Name for the file where graph is saved. Default is dependency-graph.md`")
+    @get:Option(
+        option = "graphFileName",
+        description = "Name for the file where graph is saved. Default is dependency-graph.md`",
+    )
     @get:Optional
     abstract val graphFileName: Property<String>
 
     @get:Input
-    @get:Option(option = "graphDirection", description = "The direction in which dependency graph should be laid out. Default is LR.")
+    @get:Option(
+        option = "graphDirection",
+        description = "The direction in which dependency graph should be laid out. Default is LR.",
+    )
     @get:Optional
     abstract val graphDirection: Property<Direction>
 
@@ -111,342 +116,47 @@ abstract class DependencyGraphTask : DefaultTask() {
         // Create graph of all dependencies
         val graph = graphDetails.get()
 
-        // For each module, draw its sub graph of dependencies and dependents
-        graph.projects.forEach { drawDependencies(it, graph, false, project.rootDir) }
+        val moduleBaseUrl = createGraphUrl(
+            repoUrl = repoRootUrl.orNull,
+            mainBranchName = mainBranchName.orNull,
+        )
+
+        // Draw sub graph of dependencies and dependents for each module
+        graph.projects.forEach {
+            drawDependencies(
+                currentProject = it,
+                graphDetails = graph,
+                isRootGraph = false,
+                rootDir = project.rootDir,
+                moduleBaseUrl = moduleBaseUrl,
+            )
+        }
 
         // Draw the full graph of all modules
-        drawDependencies(project.rootProject.asModuleProject(), graph, true, project.rootDir)
+        drawDependencies(
+            currentProject = project.rootProject.asModuleProject(),
+            graphDetails = graph,
+            isRootGraph = true,
+            rootDir = project.rootDir,
+            moduleBaseUrl = moduleBaseUrl,
+        )
     }
 
-    /**
-     * Returns a list of all modules that are direct or indirect dependencies of the provided module
-     * @param currentProjectAndDependencies the module(s) whose dependencies we need
-     * @param dependencies hash map of dependencies generated by [createGraph]
-     * @return List of module and all its direct & indirect dependencies
-     */
-    private fun gatherDependencies(
-        currentProjectAndDependencies: MutableList<ModuleProject>,
-        dependencies: LinkedHashMap<DependencyPair, List<String>>,
-    ): MutableList<ModuleProject> {
-        var addedNew = false
-        dependencies
-            .map { it.key }
-            .forEach { (currProject, dependencyProject) ->
-                if (
-                    currentProjectAndDependencies.contains(currProject) &&
-                    !currentProjectAndDependencies.contains(dependencyProject)
-                ) {
-                    currentProjectAndDependencies.add(dependencyProject)
-                    addedNew = true
-                }
-            }
-        return if (addedNew) {
-            gatherDependencies(
-                currentProjectAndDependencies = currentProjectAndDependencies,
-                dependencies = dependencies,
-            )
+    private fun createGraphUrl(repoUrl: String?, mainBranchName: String?): String? {
+        if (repoUrl == null) {
+            return null
+        }
+
+        val branchName = mainBranchName ?: DEFAULT_BRANCH_NAME
+
+        return if (repoUrl.endsWith('/')) {
+            "${repoUrl}blob/$branchName"
         } else {
-            currentProjectAndDependencies
+            "${repoUrl}/blob/$branchName"
         }
     }
 
-    /**
-     * Returns a list of all modules that depend on the given module
-     * @param currentProject the module whose dependencies we need
-     * @param dependencies hash map of dependencies generated by [createGraph]
-     * @return List of all modules that depend on the given module
-     */
-    private fun gatherDependents(
-        currentProject: ModuleProject,
-        dependencies: LinkedHashMap<DependencyPair, List<String>>,
-    ): List<ModuleProject> {
-        return dependencies
-            .filter { (key, _) ->
-                key.target == currentProject
-            }
-            .map { (key, _) -> key.origin }
-    }
-
-    @Suppress("LongMethod", "CyclomaticComplexMethod")
-    /**
-     * Creates a graph of dependencies for the given project and writes it to a file in the project's
-     * directory.
-     */
-    private fun drawDependencies(
-        currentProject: ModuleProject,
-        graphDetails: GraphDetails,
-        isRootGraph: Boolean,
-        rootDir: File,
-    ) {
-        val projects: LinkedHashSet<ModuleProject> = graphDetails.projects
-        val dependencies: LinkedHashMap<DependencyPair, List<String>> =
-            graphDetails.dependencies
-        val multiplatformProjects = graphDetails.multiplatformProjects
-        val androidProjects = graphDetails.androidProjects
-        val javaProjects = graphDetails.javaProjects
-        val rootProjects = graphDetails.rootProjects
-
-        val currentProjectDependencies =
-            gatherDependencies(mutableListOf(currentProject), dependencies)
-        val dependents = gatherDependents(currentProject, dependencies)
-
-        var fileText = """
-    ```mermaid
-    %%{ init: { 'theme': 'base' } }%%
-    graph LR;
-
-    %% Styling for module nodes by type
-    classDef rootNode stroke-width:4px;
-    classDef mppNode fill:#ffd2b3;
-    classDef andNode fill:#baffc9;
-    classDef javaNode fill:#ffb3ba;
-
-    %% Graph types
-    subgraph Legend
-      direction TB;
-      rootNode[Root/current module]:::rootNode;
-      andNode([Android]):::andNode;
-      javaNode{{Java}}:::javaNode;
-      mppNode([Multi-platform]):::mppNode;
-    end
-
-    %% Modules
-
-        """.trimIndent()
-        // This ensures the graph is wrapped in a box with a background, so it's consistently visible
-        // when rendered in dark mode.
-        fileText += "subgraph  \n  direction LR;\n"
-
-        val normalNodeStart = "(["
-        val normalNodeEnd = "])"
-        val rootNodeStart = "["
-        val rootNodeEnd = "]"
-        val javaNodeStart = "{{"
-        val javaNodeEnd = "}}"
-
-        var clickText = ""
-
-        for (project in projects) {
-            if (
-                !isRootGraph &&
-                !(currentProjectDependencies.contains(project) || dependents.contains(project))
-            ) {
-                continue
-            }
-            val isRoot = if (isRootGraph) {
-                rootProjects.contains(project) || project == currentProject
-            } else {
-                project == currentProject
-            }
-
-            var nodeStart = if (isRoot) {
-                rootNodeStart
-            } else {
-                normalNodeStart
-            }
-            var nodeEnd = if (isRoot) {
-                rootNodeEnd
-            } else {
-                normalNodeEnd
-            }
-
-            val nodeClass = if (multiplatformProjects.contains(project)) {
-                ":::mppNode"
-            } else if (androidProjects.contains(project)) {
-                ":::andNode"
-            } else if (javaProjects.contains(project)) {
-                if (!isRoot) {
-                    nodeStart = javaNodeStart
-                    nodeEnd = javaNodeEnd
-                }
-                ":::javaNode"
-            } else {
-                ""
-            }
-
-            fileText += "  ${project.path}${nodeStart}${project.path}${nodeEnd}$nodeClass;\n"
-
-            val relativePath = project.projectDir.relativeTo(rootDir)
-            clickText += "click ${project.path} ${GraphDetails.RepoPath}/${relativePath}\n"
-        }
-
-        fileText += """
-    end
-
-    %% Dependencies
-
-        """.trimIndent()
-
-        dependencies
-            .filter { (key, _) ->
-                val (origin, target) = key
-                (isRootGraph || currentProjectDependencies.contains(origin)) &&
-                    origin.path != target.path
-            }
-            .forEach { (key, traits) ->
-                val (origin, target) = key
-                val isApi = traits.isNotEmpty() && traits[0] == "api"
-                val isDirectDependency = origin == currentProject
-
-                val arrow = when {
-                    isApi && isDirectDependency -> "==API===>"
-                    isApi -> "--API--->"
-                    isDirectDependency -> "===>"
-                    else -> "--->"
-                }
-                fileText += "${origin.path}${arrow}${target.path}\n"
-            }
-
-        fileText += """
-
-        %% Dependents
-
-        """.trimIndent()
-
-        dependencies
-            .filter { (key, _) ->
-                val (origin, target) = key
-                dependents.contains(origin) &&
-                    target == currentProject &&
-                    origin.path != target.path
-            }
-            .forEach { (key, traits) ->
-                val (origin, target) = key
-                // bold dashed arrows aren't supported
-                val isApi = traits.isNotEmpty() && traits[0] == "api"
-                val arrow = if (isApi) {
-                    "-.API.->"
-                } else {
-                    "-.->"
-                }
-                fileText += "${origin.path}${arrow}${target.path}\n"
-            }
-
-        fileText += """
-
-%% Click interactions
-$clickText```
-""".trimIndent()
-
-        val graphFile = File(currentProject.projectDir, GraphDetails.GraphFileName)
-        graphFile.parentFile.mkdirs()
-        graphFile.delete()
-        graphFile.writeText(fileText)
-
-        println("Project module dependency graph created at ${graphFile.absolutePath}")
+    companion object {
+        private const val DEFAULT_BRANCH_NAME = "main"
     }
 }
-
-@Suppress("LongMethod", "CyclomaticComplexMethod")
-/**
- * Create a graph of all project modules, their types, dependencies and root projects.
- * @return An object of type GraphDetails containing all details
- */
-internal fun createGraph(rootProject: Project): GraphDetails {
-    val rootProjects = mutableListOf<Project>()
-    var queue = mutableListOf(rootProject)
-
-    // Traverse the list of all subfolders starting with root project and add them to
-    // rootProjects
-    while (queue.isNotEmpty()) {
-        val project = queue.removeAt(0)
-        if (project.name != GraphDetails.SystemTestName) {
-            rootProjects.add(project)
-        }
-        queue.addAll(project.childProjects.values)
-    }
-
-    val projects = LinkedHashSet<Project>()
-    val dependencies = LinkedHashMap<DependencyPair, List<String>>()
-    val multiplatformProjects = mutableListOf<Project>()
-    val androidProjects = mutableListOf<Project>()
-    val javaProjects = mutableListOf<Project>()
-
-    // Again traverse the list of all subfolders starting with the current project
-    // * Add projects to project-type lists
-    // * Add project dependencies to dependency hashmap with record for api/impl
-    // * Add projects & their dependencies to projects list
-    // * Remove any dependencies from rootProjects list
-    queue = mutableListOf(rootProject)
-    while (queue.isNotEmpty()) {
-        val project = queue.removeAt(0)
-        if (project.name == GraphDetails.SystemTestName) {
-            continue
-        }
-        queue.addAll(project.childProjects.values)
-
-        if (project.plugins.hasPlugin("org.jetbrains.kotlin.multiplatform")) {
-            multiplatformProjects.add(project)
-        }
-        if (
-            project.plugins.hasPlugin("com.android.library") ||
-            project.plugins.hasPlugin("com.android.application")
-        ) {
-            androidProjects.add(project)
-        }
-        if (
-            project.plugins.hasPlugin("java-library") ||
-            project.plugins.hasPlugin("java") ||
-            project.plugins.hasPlugin("org.jetbrains.kotlin.jvm")
-        ) {
-            javaProjects.add(project)
-        }
-
-        project.configurations.all { config ->
-            config.dependencies
-                .filterIsInstance(ProjectDependency::class.java)
-                .map { it.dependencyProject }
-                .forEach { dependency ->
-                    projects.add(project)
-                    projects.add(dependency)
-                    if (
-                        project.name != GraphDetails.SystemTestName &&
-                        project.path != dependency.path
-                    ) {
-                        rootProjects.remove(dependency)
-                    }
-
-                    val graphKey =
-                        DependencyPair(project.asModuleProject(), dependency.asModuleProject())
-                    val traits = dependencies
-                        .computeIfAbsent(graphKey) { mutableListOf() } as MutableList
-
-                    if (config.name.lowercase(Locale.getDefault()).endsWith("implementation")) {
-                        traits.add("impl")
-                    } else {
-                        traits.add("api")
-                    }
-                }
-        }
-    }
-
-    // Collect leaf projects which may be denoted with a different shape or rank
-    val leafProjects = mutableListOf<Project>()
-    projects.forEach {
-        val allDependencies = it.configurations
-            .map { config ->
-                config.dependencies
-                    .filterIsInstance(ProjectDependency::class.java)
-                    .filter { dependency ->
-                        dependency.dependencyProject.path != it.path
-                    }
-            }
-
-        if (allDependencies.isEmpty()) {
-            leafProjects.add(it)
-        } else {
-            leafProjects.remove(it)
-        }
-    }
-
-    return GraphDetails(
-        projects = LinkedHashSet(projects.map { it.asModuleProject() }.sortedBy { it.path }),
-        dependencies = dependencies,
-        multiplatformProjects = multiplatformProjects.map { it.asModuleProject() },
-        androidProjects = androidProjects.map { it.asModuleProject() },
-        javaProjects = javaProjects.map { it.asModuleProject() },
-        rootProjects = rootProjects.map { it.asModuleProject() },
-    )
-}
-
-internal fun Project.asModuleProject() = ModuleProject(this.path, this.projectDir)
